--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,2 @@
+node_modules
+bench/src/jquery.js