--- a/.npmignore
+++ b/.npmignore
@@ -1,0 +1,1 @@
+bench/src/jquery.js