@substack I'd like your approval on this - it'd be great for node-detective and insert-module-globals.

---

Before:

```
$ for i in {1..5}; do node bench/walk.js; done
116ms - 35066
118ms - 35066
131ms - 35066
127ms - 35066
119ms - 35066
```

After:

```
$ for i in {1..5}; do node bench/walk.js; done
80ms - 35066
83ms - 35066
82ms - 35066
81ms - 35066
81ms - 35066
```